feat: verification model, send email verification
