+++ src/main/kotlin/changchanghwang/parrotworldserver/common/ddd/Aggregate.kt
@@ -1,0 +1,18 @@
+package changchanghwang.parrotworldserver.common.ddd
+
+import jakarta.persistence.EntityListeners
+import jakarta.persistence.MappedSuperclass
+import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+
+@EntityListeners(AuditingEntityListener::class)
+@MappedSuperclass
+open class Aggregate {
+    @CreatedDate
+    var createdAt: LocalDateTime? = null
+
+    @LastModifiedDate
+    var updatedAt: LocalDateTime? = null
+}
