--- src/main/kotlin/changchanghwang/parrotworldserver/members/domain/MemberRole.kt
@@ -1,6 +1,0 @@
-package changchanghwang.parrotworldserver.members.domain
-
-enum class MemberRole {
-    USER,
-    ADMIN,
-}
