+++ src/main/kotlin/changchanghwang/parrotworldserver/services/members/domain/MemberRole.kt
@@ -1,0 +1,6 @@
+package changchanghwang.parrotworldserver.services.members.domain
+
+enum class MemberRole(val value: String) {
+    USER("USER"),
+    ADMIN("ADMIN"),
+}
