+++ src/main/kotlin/changchanghwang/parrotworldserver/services/members/infrastructure/MemberJpaRepository.kt
@@ -1,0 +1,12 @@
+package changchanghwang.parrotworldserver.services.members.infrastructure
+
+import changchanghwang.parrotworldserver.services.members.domain.Member
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface MemberJpaRepository : JpaRepository<Member, Long> {
+    fun findByEmail(email: String): Member?
+
+    fun findByNickName(nickName: String): Member?
+
+    fun findByRefreshToken(refreshToken: String): Member?
+}
