+++ src/main/kotlin/changchanghwang/parrotworldserver/services/verifcation/domain/VerificationType.kt
@@ -1,0 +1,5 @@
+package changchanghwang.parrotworldserver.services.verifcation.domain
+
+enum class VerificationType {
+    SIGNUP,
+}
