+++ src/main/kotlin/changchanghwang/parrotworldserver/services/verifcation/infrastructure/VerificationJpaRepository.kt
@@ -1,0 +1,6 @@
+package changchanghwang.parrotworldserver.services.verifcation.infrastructure
+
+import changchanghwang.parrotworldserver.services.verifcation.domain.Verification
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface VerificationJpaRepository : JpaRepository<Verification, Long>
