+++ src/main/kotlin/changchanghwang/parrotworldserver/services/verifcation/infrastructure/VerificationRepository.kt
@@ -1,0 +1,7 @@
+package changchanghwang.parrotworldserver.services.verifcation.infrastructure
+
+import changchanghwang.parrotworldserver.services.verifcation.domain.Verification
+
+interface VerificationRepository {
+    fun save(verification: Verification): Verification
+}
