+++ src/main/resources/templates/sign-up-validation-email.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="kor" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+    <div style="max-width:600px; margin:auto">
+        <div style="margin:auto; display:flex; flex-direction:column; gap:32px;">
+            <img alt="parrot-world" style="aspect-ratio:auto 1463 / 574; width:200px;" src="https://parrot-world-img.s3.ap-northeast-2.amazonaws.com/logo.png" />
+            <p style="font-size:26px; font-weight: bold;">
+                이메일 인증을 진행해주세요
+            </p>
+            <div>
+                <p>
+                    앵무새 세상을 이용해주셔서 감사합니다 :)
+                </p>
+                <p>
+                    가입을 위해 아래 인증번호를 화면에 입력해주세요.
+                </p>
+            </div>
+            <div style="height:5rem;background-color:#E3E3FF;text-align:center; vertical-align:middle;">
+                <span style="font-size: 3rem; font-weight: bold; letter-spacing: 8px;">
+                   [[${verificationCode}]]
+                </span>
+            </div>
+            <span>
+                본 인증번호의 유효기간은 3분 입니다.
+            </span>
+        </div>
+    </div>
+</body>
+</html>
